Moved entire database module to core
