+++ app/src/main/java/com/yashovardhan99/healersdiary/create/ChoosePatientAdapter.kt
@@ -4,7 +4,7 @@ import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
-import com.yashovardhan99.healersdiary.database.Patient
+import com.yashovardhan99.core.database.Patient
 import com.yashovardhan99.healersdiary.databinding.ItemChoosePatientBinding
 import com.yashovardhan99.healersdiary.patients.PatientDiff
 import timber.log.Timber
