+++ app/src/main/java/com/yashovardhan99/healersdiary/create/CreateActivityViewModel.kt
@@ -3,10 +3,10 @@ package com.yashovardhan99.healersdiary.create
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.yashovardhan99.healersdiary.dashboard.DashboardRepository
-import com.yashovardhan99.healersdiary.database.ActivityType
-import com.yashovardhan99.healersdiary.database.Healing
-import com.yashovardhan99.healersdiary.database.Patient
-import com.yashovardhan99.healersdiary.database.Payment
+import com.yashovardhan99.core.database.ActivityType
+import com.yashovardhan99.core.database.Healing
+import com.yashovardhan99.core.database.Patient
+import com.yashovardhan99.core.database.Payment
 import com.yashovardhan99.healersdiary.utils.Request
 import com.yashovardhan99.core.setToStartOfDay
 import dagger.hilt.android.lifecycle.HiltViewModel
