+++ app/src/main/java/com/yashovardhan99/healersdiary/onboarding/OnboardingRepository.kt
@@ -1,6 +1,6 @@
 package com.yashovardhan99.healersdiary.onboarding
 
-import com.yashovardhan99.healersdiary.database.HealersDao
+import com.yashovardhan99.core.database.HealersDao
 import com.yashovardhan99.core.DangerousDatabase
 import javax.inject.Inject
 import javax.inject.Singleton
