+++ app/src/main/java/com/yashovardhan99/healersdiary/onboarding/OnboardingViewModel.kt
@@ -9,7 +9,7 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.yashovardhan99.healersdiary.AppDataStore
+import com.yashovardhan99.core.AppDataStore
 import com.yashovardhan99.core.analytics.AnalyticsEvent
 import com.yashovardhan99.core.DangerousDatabase
 import dagger.hilt.android.lifecycle.HiltViewModel
