+++ app/src/main/java/com/yashovardhan99/healersdiary/patients/PatientDetailRepository.kt
@@ -3,10 +3,10 @@ package com.yashovardhan99.healersdiary.patients
 import androidx.paging.Pager
 import androidx.paging.PagingConfig
 import androidx.paging.PagingData
-import com.yashovardhan99.healersdiary.database.HealersDao
-import com.yashovardhan99.healersdiary.database.Healing
-import com.yashovardhan99.healersdiary.database.Patient
-import com.yashovardhan99.healersdiary.database.Payment
+import com.yashovardhan99.core.database.HealersDao
+import com.yashovardhan99.core.database.Healing
+import com.yashovardhan99.core.database.Patient
+import com.yashovardhan99.core.database.Payment
 import kotlinx.coroutines.flow.Flow
 import java.util.*
 import javax.inject.Inject
