+++ app/src/main/java/com/yashovardhan99/healersdiary/patients/PatientDetailViewModel.kt
@@ -11,7 +11,7 @@ import com.yashovardhan99.core.setToStartOfDay
 import com.yashovardhan99.core.setToStartOfLastMonth
 import com.yashovardhan99.core.setToStartOfMonth
 import com.yashovardhan99.healersdiary.create.CreateRepository
-import com.yashovardhan99.healersdiary.database.*
+import com.yashovardhan99.core.database.*
 import com.yashovardhan99.healersdiary.utils.*
 import com.yashovardhan99.healersdiary.utils.HealingParent.Healing.Companion.toUiHealing
 import com.yashovardhan99.healersdiary.utils.PaymentParent.Payment.Companion.toUiPayment
