+++ app/src/main/java/com/yashovardhan99/healersdiary/patients/PatientsListFragment.kt
@@ -18,7 +18,7 @@ import com.yashovardhan99.core.analytics.AnalyticsEvent
 import com.yashovardhan99.core.transitionDurationLarge
 import com.yashovardhan99.healersdiary.R
 import com.yashovardhan99.healersdiary.dashboard.DashboardViewModel
-import com.yashovardhan99.healersdiary.database.Patient
+import com.yashovardhan99.core.database.Patient
 import com.yashovardhan99.healersdiary.databinding.FragmentPatientsListBinding
 import com.yashovardhan99.healersdiary.utils.*
 import dagger.hilt.android.AndroidEntryPoint
