+++ app/src/main/java/com/yashovardhan99/healersdiary/utils/Activity.kt
@@ -2,7 +2,7 @@ package com.yashovardhan99.healersdiary.utils
 
 import androidx.annotation.StringRes
 import com.yashovardhan99.healersdiary.R
-import com.yashovardhan99.healersdiary.database.Patient
+import com.yashovardhan99.core.database.Patient
 import java.util.*
 
 sealed class ActivityParent {
