+++ app/src/main/java/com/yashovardhan99/healersdiary/utils/Stat.kt
@@ -2,7 +2,7 @@ package com.yashovardhan99.healersdiary.utils
 
 import androidx.annotation.StringRes
 import com.yashovardhan99.healersdiary.R
-import com.yashovardhan99.healersdiary.database.ActivityType
+import com.yashovardhan99.core.database.ActivityType
 import java.math.BigDecimal
 import java.text.NumberFormat
