+++ core/src/main/java/com/yashovardhan99/core/database/HealersDatabase.kt
@@ -1,4 +1,4 @@
-package com.yashovardhan99.healersdiary.database
+package com.yashovardhan99.core.database
 
 import androidx.room.Database
 import androidx.room.RoomDatabase
