+++ core/src/main/java/com/yashovardhan99/core/database/Payment.kt
@@ -1,4 +1,4 @@
-package com.yashovardhan99.healersdiary.database
+package com.yashovardhan99.core.database
 
 import android.os.Bundle
 import androidx.core.os.bundleOf
