+++ online/src/main/java/com/yashovardhan99/healersdiary/online/OnlineComponent.kt
@@ -1,7 +1,7 @@
 package com.yashovardhan99.healersdiary.online
 
 import android.content.Context
-import com.yashovardhan99.healersdiary.OnlineModuleDependencies
+import com.yashovardhan99.core.OnlineModuleDependencies
 import com.yashovardhan99.healersdiary.online.importFirebase.ImportFirebaseFragment
 import dagger.BindsInstance
 import dagger.Component
